requirements.txt
flask
pyyaml
python-slugify
requests
pywebview
gunicorn